+++ src/main/java/com/nscharrenberg/um/multiagentsurveillance/agents/shared/algorithms/pathfinding/AStar/AStar.java
@@ -6,7 +6,7 @@ import com.nscharrenberg.um.multiagentsurveillance.agents.shared.algorithms.stru
 import com.nscharrenberg.um.multiagentsurveillance.agents.shared.algorithms.structures.UnknownAreaCalculator.UnknownAreaCalculate;
 import com.nscharrenberg.um.multiagentsurveillance.agents.shared.utils.QueueNode;
 import com.nscharrenberg.um.multiagentsurveillance.agents.shared.utils.TreeNode;
-import com.nscharrenberg.um.multiagentsurveillance.headless.models.Angle;
+import com.nscharrenberg.um.multiagentsurveillance.headless.models.Action;
 import com.nscharrenberg.um.multiagentsurveillance.headless.models.Area;
 import com.nscharrenberg.um.multiagentsurveillance.headless.models.Player;
 import com.nscharrenberg.um.multiagentsurveillance.headless.models.Tile;
@@ -54,8 +54,8 @@ public class AStar implements IPathFinding {
         board.add(player.getTile());
 
         while (!tree.getTile().equals(target)) {
-            for (Angle angle : Angle.values()) {
-                Optional<Tile> nextTileOpt = BoardUtils.nextPosition(board, tree.getTile(), angle);
+            for (Action action : Action.values()) {
+                Optional<Tile> nextTileOpt = BoardUtils.nextPosition(board, tree.getTile(), action);
 
 
                 if (nextTileOpt.isPresent() && !nextTileOpt.get().isCollision() && visited.get(nextTileOpt.get().getX()).get(nextTileOpt.get().getY()).equals(Boolean.FALSE)) {
@@ -67,11 +67,11 @@ public class AStar implements IPathFinding {
 
                     int distance = computeDistance(nextTileOpt.get(), target);
 
-                    TreeNode childNode = new TreeNode(nextTileOpt.get(), angle, tree);
+                    TreeNode childNode = new TreeNode(nextTileOpt.get(), action, tree);
 
 
                     if (!tree.getEntrancePosition().equals(childNode.getEntrancePosition())) {
-                        TreeNode additionalChildNode = new TreeNode(nextTileOpt.get(), angle, childNode);
+                        TreeNode additionalChildNode = new TreeNode(nextTileOpt.get(), action, childNode);
 
                         heap.insert(new Node(distance+0.5, additionalChildNode));
                     } else {
@@ -93,7 +93,7 @@ public class AStar implements IPathFinding {
 
 
 
-        LinkedList<Angle> sequenceMoves = new LinkedList<>();
+        LinkedList<Action> sequenceMoves = new LinkedList<>();
 
         TreeNode lastMove = tree;
