+++ src/main/java/com/nscharrenberg/um/multiagentsurveillance/agents/shared/algorithms/structures/UnknownAreaCalculator/UnknownAreaCalculate.java
@@ -1,14 +1,12 @@
 package com.nscharrenberg.um.multiagentsurveillance.agents.shared.algorithms.structures.UnknownAreaCalculator;
 
 import com.nscharrenberg.um.multiagentsurveillance.headless.models.AdvancedAngle;
-import com.nscharrenberg.um.multiagentsurveillance.headless.models.Angle;
 import com.nscharrenberg.um.multiagentsurveillance.headless.models.Area;
 import com.nscharrenberg.um.multiagentsurveillance.headless.models.Tile;
 import com.nscharrenberg.um.multiagentsurveillance.headless.utils.BoardUtils;
 
 import java.util.HashMap;
 import java.util.Map;
-import java.util.Optional;
 
 public class UnknownAreaCalculate {
