+++ src/main/java/com/nscharrenberg/um/multiagentsurveillance/agents/shared/utils/TreeNode.java
@@ -1,15 +1,12 @@
 package com.nscharrenberg.um.multiagentsurveillance.agents.shared.utils;
 
-import com.nscharrenberg.um.multiagentsurveillance.headless.models.Angle;
+import com.nscharrenberg.um.multiagentsurveillance.headless.models.Action;
 import com.nscharrenberg.um.multiagentsurveillance.headless.models.Tile;
-
-import java.util.LinkedList;
-import java.util.Queue;
 
 public class TreeNode extends QueueNode {
     private TreeNode parent;
 
-    public TreeNode(Tile tile, Angle entrancePosition, TreeNode parent) {
+    public TreeNode(Tile tile, Action entrancePosition, TreeNode parent) {
         super(tile, entrancePosition);
         this.parent = parent;
     }
