--- src/main/java/com/nscharrenberg/um/multiagentsurveillance/gui/canvas/CanvasLauncher.java
@@ -1,12 +1,0 @@
-package com.nscharrenberg.um.multiagentsurveillance.gui.canvas;
-
-public class CanvasLauncher {
-
-    public static void main(String[] args) {
-        try {
-            CanvasApp.main(args);
-        } catch (Exception ex) {
-            ex.printStackTrace();
-        }
-    }
-}
