+++ src/main/java/com/nscharrenberg/um/multiagentsurveillance/headless/models/Guard.java
@@ -1,7 +1,7 @@
 package com.nscharrenberg.um.multiagentsurveillance.headless.models;
 
 public class Guard extends Player {
-    public Guard(Tile position, Angle direction) {
+    public Guard(Tile position, Action direction) {
         // TODO: Read speed from Configuration
         super(position, direction, 10);
     }
