+++ src/main/java/com/nscharrenberg/um/multiagentsurveillance/headless/models/gui/assets/style.css
