+++ src/main/java/com/nscharrenberg/um/multiagentsurveillance/headless/models/gui/canvas/CanvasLauncher.java
@@ -1,0 +1,12 @@
+package com.nscharrenberg.um.multiagentsurveillance.headless.models.gui.canvas;
+
+public class CanvasLauncher {
+
+    public static void main(String[] args) {
+        try {
+            CanvasApp.main(args);
+        } catch (Exception ex) {
+            ex.printStackTrace();
+        }
+    }
+}
