+++ src/main/java/com/nscharrenberg/um/multiagentsurveillance/headless/models/gui/javafx/controllers/HomeScreen.java
@@ -1,8 +1,6 @@
-package com.nscharrenberg.um.multiagentsurveillance.gui.javafx.controllers;
+package com.nscharrenberg.um.multiagentsurveillance.headless.models.gui.javafx.controllers;
 
 import com.nscharrenberg.um.multiagentsurveillance.Application;
-import javafx.event.ActionEvent;
-import javafx.event.EventHandler;
 import javafx.geometry.Pos;
 import javafx.scene.Scene;
 import javafx.scene.control.Button;
