+++ src/main/java/com/nscharrenberg/um/multiagentsurveillance/headless/models/gui/terminal/TerminalApp.java
@@ -1,4 +1,4 @@
-package com.nscharrenberg.um.multiagentsurveillance.gui.terminal;
+package com.nscharrenberg.um.multiagentsurveillance.headless.models.gui.terminal;
 
 public class TerminalApp {
     public static void main(String[] args) throws Exception {
