+++ src/main/java/com/nscharrenberg/um/multiagentsurveillance/headless/utils/AngleConverter.java
@@ -1,31 +1,31 @@
 package com.nscharrenberg.um.multiagentsurveillance.headless.utils;
 
-import com.nscharrenberg.um.multiagentsurveillance.headless.models.Angle;
+import com.nscharrenberg.um.multiagentsurveillance.headless.models.Action;
 
 public class AngleConverter {
 
     /**
-     * Converting an angle to a predefined Angle field.
+     * Converting an angle to a predefined Action field.
      * @param angle - angle in degrees to be converted
-     * @return - predefined Angle {UP, DOWN, LEFT, RIGHT}
+     * @return - predefined Action {UP, DOWN, LEFT, RIGHT}
      */
-    public static Angle convert(double angle) {
-        // Right when Angle is between 45 and 135 degrees
+    public static Action convert(double angle) {
+        // Right when Action is between 45 and 135 degrees
         if (angle > 45 && angle <= 135) {
-            return Angle.RIGHT;
+            return Action.RIGHT;
         }
 
-        // Down when Angle is between 135 and 225 degrees
+        // Down when Action is between 135 and 225 degrees
         if (angle > 135 && angle <= 225) {
-            return Angle.DOWN;
+            return Action.DOWN;
         }
 
-        // Left when Angle is between 225 and 315 degrees
+        // Left when Action is between 225 and 315 degrees
         if (angle > 225 && angle <= 315) {
-            return Angle.LEFT;
+            return Action.LEFT;
         }
 
-        // UP when Angle is between -45 (315) and 45 degrees
-        return Angle.UP;
+        // UP when Action is between -45 (315) and 45 degrees
+        return Action.UP;
     }
 }
