+++ src/test/java/headless/MapBuildTest.java
@@ -1,7 +1,5 @@
 package headless;
 
-import com.nscharrenberg.um.multiagentsurveillance.agents.random.RandomAgent;
-import com.nscharrenberg.um.multiagentsurveillance.agents.shared.Agent;
 import com.nscharrenberg.um.multiagentsurveillance.headless.Factory;
 import com.nscharrenberg.um.multiagentsurveillance.headless.exceptions.*;
 import com.nscharrenberg.um.multiagentsurveillance.headless.models.*;
@@ -10,7 +8,6 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import java.util.List;
 import java.util.Optional;
 
 public class MapBuildTest {
@@ -108,7 +105,7 @@ public class MapBuildTest {
 
         Tile tile = tileOpt.get();
 
-        Guard guard = new Guard(tile, Angle.UP);
+        Guard guard = new Guard(tile, Action.UP);
         tile.add(guard);
 
         // Check if player exists
@@ -116,7 +113,7 @@ public class MapBuildTest {
         Assertions.assertEquals(1, tile.getItems().size());
         Assertions.assertEquals(guard, tile.getItems().get(0));
 
-        Angle move = Angle.DOWN;
+        Action move = Action.DOWN;
 
         // Rotate the player
         Factory.getPlayerRepository().move(guard, move);
@@ -160,7 +157,7 @@ public class MapBuildTest {
 
         Tile tile = tileOpt.get();
 
-        Guard guard = new Guard(tile, Angle.UP);
+        Guard guard = new Guard(tile, Action.UP);
         tile.add(guard);
 
         // Check if player exists
@@ -168,7 +165,7 @@ public class MapBuildTest {
         Assertions.assertEquals(1, tile.getItems().size());
         Assertions.assertEquals(guard, tile.getItems().get(0));
 
-        Angle move = Angle.UP;
+        Action move = Action.UP;
 
         // build wall
         int wallX = tile.getX() + move.getxIncrement();
@@ -197,7 +194,7 @@ public class MapBuildTest {
         int x2 = 6;
         int y2 = 3;
 
-        Angle facingDirection = Angle.LEFT;
+        Action facingDirection = Action.LEFT;
 
         // Destination
         int destX = 20;
@@ -217,10 +214,10 @@ public class MapBuildTest {
 
         Tile tile = tileOpt.get();
 
-        Guard guard = new Guard(tile, Angle.RIGHT);
+        Guard guard = new Guard(tile, Action.RIGHT);
         tile.add(guard);
 
-        Angle move = Angle.RIGHT;
+        Action move = Action.RIGHT;
 
         // build wall
         int teleportSourceX = tile.getX() + move.getxIncrement();
